:art: 잘못된 디렉토리 수정
